+++ packages/deepsea-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # deepsea-components
 
+## 5.17.31
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.47.7
+    - soda-hooks@6.16.20
+
 ## 5.17.30
 
 ### Patch Changes